# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dtos.LoginResponse:
    properties:
      access_token:
        type: string
      active_tenant_id:
        type: string
      expires_in:
        description: access token lifetime (seconds)
        type: integer
      refresh_expires_in:
        description: refresh token lifetime (seconds)
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
    type: object
...
  dtos.Meta:
    description: Metadata for pagination
    properties:
      code:
        example: 400
        type: integer
      error_code:
        example: BAD_REQUEST
        type: string
      message:
        example: Request is invalid
        type: string
      page:
        example: 1
        type: integer
      page_size:
        example: 20
        type: integer
      total:
        example: 18
        type: integer
    type: object
...
  dtos.WebauthnCredentialResponse:
    properties:
      created_at:
        type: string
      device_name:
        example: Work MacBook
        type: string
      id:
        example: 00000000-0000-0000-0000-000000000001
        type: string
    type: object
  dtos.WebauthnLoginFinishRequest:
    properties:
      credential:
        description: Credential is the browser PublicKeyCredential JSON from navigator.credentials.get().
        type: object
      email:
        example: user@example.com
        type: string
      remember_me:
        example: false
        type: boolean
      return_to:
        example: ""
        type: string
      session_token:
        example: 9b2c4e8a-1d3f-4a5b-8c7e-0f1a2b3c4d5e
        type: string
      tenant_id:
        example: 00000000-0000-0000-0000-000000000001
        type: string
    required:
    - email
    - session_token
    type: object
  dtos.WebauthnLoginStartRequest:
    properties:
      email:
        example: user@example.com
        type: string
      tenant_id:
        example: 00000000-0000-0000-0000-000000000001
        type: string
    required:
    - email
    type: object
  dtos.WebauthnLoginStartResponse:
    properties:
      options:
        type: object
      session_token:
        example: 9b2c4e8a-1d3f-4a5b-8c7e-0f1a2b3c4d5e
        type: string
    type: object
  dtos.WebauthnRegisterFinishRequest:
    properties:
      credential:
        description: Credential is the browser PublicKeyCredential JSON from navigator.credentials.create().
        type: object
      session_token:
        example: 9b2c4e8a-1d3f-4a5b-8c7e-0f1a2b3c4d5e
        type: string
    required:
    - session_token
    type: object
  dtos.WebauthnRegisterStartRequest:
    properties:
      device_name:
        description: DeviceName optional label stored with the passkey (e.g. "Work
          MacBook").
        example: Personal iPhone
        maxLength: 255
        type: string
    type: object
  dtos.WebauthnRegisterStartResponse:
    properties:
      options:
        type: object
      session_token:
        example: 9b2c4e8a-1d3f-4a5b-8c7e-0f1a2b3c4d5e
        type: string
    type: object
info:
  contact: {}
  description: GateForge IAM application API under /api/v1. The canonical public SDK
    contract (including root OIDC paths) lives at api/openapi.yaml in the monorepo.
  title: GateForge IAM API
  version: 0.1.0
paths:
...
  /webauthn/credentials:
    get:
      description: Returns passkeys registered for the current user (device label
        and created date only).
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/dtos.WebauthnCredentialResponse'
                type: array
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      security:
      - BearerAuth: []
      summary: List registered passkeys
      tags:
      - WebAuthn
  /webauthn/login/finish:
    post:
      consumes:
      - application/json
      description: 'Verifies the assertion. On success without MFA: returns `data`
        as LoginResponse and sets `iam_session` cookie. When MFA is enabled for the
        user, returns MFALoginChallengeResponse in `data` instead (then call POST
        /mfa/challenge/verify).'
      parameters:
      - description: Assertion response
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.WebauthnLoginFinishRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'When MFA is off: LoginResponse + iam_session. When MFA is
            on: `data` is MFALoginChallengeResponse (see description).'
          schema:
            properties:
              data:
                $ref: '#/definitions/dtos.LoginResponse'
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      summary: Finish passkey login
      tags:
      - WebAuthn
  /webauthn/login/start:
    post:
      consumes:
      - application/json
      description: Returns WebAuthn PublicKeyCredentialRequestOptions as `data.options`
        and a `session_token` for login/finish. Public endpoint (rate limited).
      parameters:
      - description: User lookup
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.WebauthnLoginStartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/dtos.WebauthnLoginStartResponse'
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      summary: Begin passkey login
      tags:
      - WebAuthn
  /webauthn/register/finish:
    post:
      consumes:
      - application/json
      description: Verifies the attestation and stores the passkey for the current
        user. Body includes `session_token` from register/start and the PublicKeyCredential
        JSON from the browser.
      parameters:
      - description: Credential response
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.WebauthnRegisterFinishRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: object
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      security:
      - BearerAuth: []
      summary: Finish passkey registration
      tags:
      - WebAuthn
  /webauthn/register/start:
    post:
      consumes:
      - application/json
      description: Returns WebAuthn PublicKeyCredentialCreationOptions as `data.options`
        and a `session_token` to send with register/finish. Requires a logged-in user
        (Bearer access token).
      parameters:
      - description: Optional device label
        in: body
        name: body
        schema:
          $ref: '#/definitions/dtos.WebauthnRegisterStartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/dtos.WebauthnRegisterStartResponse'
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      security:
      - BearerAuth: []
      summary: Begin passkey registration
      tags:
      - WebAuthn
schemes:
- http
- https
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: Bearer Token Authentication. Use "Bearer {token}" as the value.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"